+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Store more variables in mongodb
 - Make more variables configurable from frontend settingspage
 
+## [v13.0.1.beta] - 28-11-10
+### Added
+- Jest and supertest to make tests available
+
+### Changed
+- Position of Snackbar
+
+### Fixed
+- Villa Anna Markise Page will now show snackbar
+
 ## [v12.3.1] - 28-11-07
 ### Changed
 - Dockerfile to meet new structure